Refactoring useReadResource, useResolveResource, useCollection
There's a lot of logic shared between the hooks, and this is a first
step towards refactoring that a bit.

Originally the plan was to have a `useReadResource` that would be used
by both `useCollection` and `useResource`. This is useful, because a lot
of features like `refreshOnStale` has to be implemented twice. If there
were a more generic, fundemental hook that both `useResource` and
`useCollection` could use, this might potentially be a lot simpler.

So this is a first step towards doing that. useReadResource now listens
for events, and has some new options that make it good enough for
`useCollection`.

All `useCollection` now does is re-expose `useReadResource` but with an
`items` result that contains all the collection items. It's much simpler
now.

In the future I hope to refactor this further so `useResource` can also
undergo this same refactoring and become a lot simpler.

Until then all new features will still have to be implemented in two
places, but now its:

* useReadResource
* useResource