usePagedCollection
This hook is similar to `useCollection`, but it lets load in additional
pages from a collection. The hook assumes that a "next" link
relationship exists on the collection.

Every time the next page is loaded, the new items from the new page are
appended to the item list. This lets users do a 'infinite scroll' type
interface.