Hooks now are now consistent.
All hooks now have the same function signature:

```typescript
useFoo(resource, options)
```

This was already the case for `useCollection` and `usePagedCollection`,
but this has also been changed for `useResource`.

This is a breaking change. All calls such as the following stay the
same:

```typescript
useResource(resource);
useResource('/res');
useResource(Promise.resolve(resource));
```

But if options were used ,this syntax is no longer valid:

```typescript
useResource({
  resource: res,
  mode: 'POST',
  initialData: 'bla',
  refreshOnStale: true,
});
```

The `resource` has been moved out of the options object. The next syntax
is:

```typescript
useResource(res, {
  mode: 'POST',
  initialData: 'bla'
  refreshOnStale: true,
});
```

Given that this is a breaking change, this will be released as 3.0