Fix white screen of death if an updated string resourceLike is passed.
When useResource is called in re-renders, with an updated value for
'resourceLike', and that value has a type string, the useResource hook
might temporarily return 'undefined' for resource, while 'loading' is
false and 'error' is null.

Normally if that component gets a new resourceLike and needs to spend
time resolving it, it should make sure loading flips to 'true'.

However, if we get resourceLike as a string we can actually skip the
loading step altogether and immediately return the new resource.

This PR:

1. Fixes the bug
2. Skips a potential re-render
3. Re-uses code and removes some complexity