Creating a 'useNewResource' hook.
This is currently not functioning, but a step in the right direction.

`useReadResource` is meant to be a robust foundation for all other hooks
(including useCollection, usePagedCollection), but it wasn't possible
yet to refactor useResource to use it.

The main reason for this is because useResource has support for 'POST'
mode, which lets a user create a new resource with POST and then
continue editing. This feature hasn't worked well for a while.

So this PR is meant to kill a few birds with one stone:

* Drastically simplifiy `useResource` (`useResourceSimple` should eventually
  replace it).
* `useResource` should use `useReadResource` under the hook, reducing
  duplication.
* Create a new `useNewResource`, specific for the `POST` use-case, and
  remove that functionality from `useResource`.

Because `useResource` can no longer create new resources after this,
this does consitute a new major resource of this package.