Renaming usePagedCollection to useInfiniteCollection
The true reason for this is that I want to introduce a new hook later
that lets user go from page to page, using previous and next buttons and
pick page numbers.

Having both that hook, and this existing one means that there's 2 hooks
that deal with pages, and we need good names for both. I feel that this
existing hook would be better if it were called useInfiniteCollection,
so this PR prepares for that change, without breaking BC.