Reduce the number of renders.
Before this change `useReadResource` rendered 5 times from a cold cache.
After this change this is 3.

Because this interal hook is used by `useCollection`,
`useInfiniteCollection`, `useNewResource`, `useResource` also all have
their number of renders reduced by 2.

In the future it will be possible to get this 3 down to 2. Alledgedly
with React 18 this will automatically happen.

The reason is that there's a few places that update 2 state variables in
a row, looking like this:

```typescript
setResourceState(true)
setLoading(false)
```

To my surprise, even if these statements happen immediately after each
other, these changes are not batched and cause 2 renders immediately
after each other.

It would also be possible to fix this by us, if we combine 'loading' and
'resourceState' into a single state variable, but this will require a
larger refactor.

With this change I also consider #64 done.

Fixes #64