Adding a <SelectLink> component.
This is a pattern I noticed I've had to do more than a few times, and it
made sense to make a default component for this.

All this does right now is render a <select> with a list of links. The
label is taken from the link title, and the uri is used for the value.

It's possible to add an empty option like this:

```typescript
<SelectLink resource="/some-collection">
  <option value="">Please select an option</option>
</SelectLink>
```